feat(in-view-video): add markup and styling
