+++ src/scripts/types/story.ts
@@ -172,6 +172,7 @@ export type Legend = {
 
 export type BaseModuleSlide = {
   url?: string;
+  track?: string;
   altText?: string;
   text: string;
   focus?: ImageFocus;
